Move new hostname functions to kano_settings.advanced. This is because other editing of the hostname file occurs there. Programmers will not need to be aware of multiple modules editing the same file.
