+++ kano_settings/system/advanced.py
@@ -17,11 +17,17 @@ import urllib2
 from bs4 import BeautifulSoup
 import pycountry
 import gzip
+import sys
+import re
+
+from kano.logging import logger
+import kano.utils
+from kano_settings.system.get_username import get_first_username
 
 from kano_settings.common import settings_dir
 from kano.utils import read_file_contents, write_file_contents, \
     read_file_contents_as_lines, read_json, write_json, ensure_dir, \
-    get_user_unsudoed
+    get_user_unsudoed, sed
 from kano.logging import logger
 from kano.network import set_dns, restore_dns_interfaces, \
     clear_dns_interfaces, refresh_resolvconf
@@ -97,6 +103,50 @@ def encrypt_password(str):
 
 def authenticate_parental_password(passwd):
     return read_file_contents(password_file) == encrypt_password(passwd)
+
+
+def set_hostname_postinst():
+    # when running as post install, get the existing first user and set as host name
+    new_hostname = get_first_username()
+
+    if new_hostname is None:
+        logger.warning("No first user")
+    else:
+        set_hostname(new_hostname)
+
+
+def set_hostname(new_hostname):
+    if os.environ['LOGNAME'] != 'root':
+        logger.error("Error: Settings must be executed with root privileges")
+
+
+    # Check username chars
+
+    new_hostname = re.sub('[^a-zA-Z0-9]', '', new_hostname).lower()
+
+    if new_hostname == '':
+        logger.error('no letters left in username after removing illegal ones')
+
+    # check if already done
+    curr_hosts = read_file_contents_as_lines('/etc/hosts')
+    if '127.0.0.1\tkano' not in curr_hosts:
+        logger.warn('/etc/hosts already modified, not changing')
+
+    try:
+        lines_changed = sed(
+                                      '127.0.0.1\s+(kano)',
+                                      '127.0.0.1\t{}'.format(new_hostname),
+                                      '/etc/hosts',
+                                      True)
+        if lines_changed != 1:
+            logger.error("failed to change /etc/hosts")
+    except:
+        logger.error("exception while changing change /etc/hosts")
+
+    try:
+        write_file_contents('/etc/hostname', new_hostname+'\n')
+    except:
+        logger.error("exception while changing change /etc/hostname")
 
 
 def create_empty_hosts():
